fix: solve warning by primary key types

+++ src/main/java/com/mechanical_workshop_usm/work_order_realized_service_module/persistence/model/WorkOrderRealizedServicesPk.java
@@ -2,5 +2,5 @@ package com.mechanical_workshop_usm.work_order_realized_service_module.persisten
 
 import java.io.Serializable;
 
-public record WorkOrderRealizedServicesPk(long workOrder, long workService) implements Serializable {
+public record WorkOrderRealizedServicesPk(int workOrder, int workService) implements Serializable {
 }
